Fix trainings API call
Commit bf1ba5e renamed the redacted API client's method `trainings` to
`employee_training_courses`, but did not update the method call from the
corresponding path of the redacted API.

Fix method call.

I've tried to set up testing for the Sinatra app, but it's turning into
a bigger thing, so in the interest of being able to debug missing events
from Productive, I'm leaving it for later development.